djchat/requirements.txt
asgiref==3.7.2
attrs==23.2.0
black==23.12.1
click==8.1.7
Django==5.0
django-filter==23.5
djangorestframework==3.14.0
drf-spectacular==0.27.0
flake8==6.1.0
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
Markdown==3.5.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.1.0
pycodestyle==2.11.1
pyflakes==3.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
rpds-py==0.16.2
sqlparse==0.4.4
uritemplate==4.1.1
